Replace size_t in shared memory by uint32_t
Eventually we may want to use uint64_t, but will need to confirm atomicity.

Bug: 12381724

+++ include/private/media/StaticAudioTrackState.h
@@ -25,9 +25,13 @@ namespace android {
 // state is wrapped by a SingleStateQueue.
 struct StaticAudioTrackState {
     // do not define constructors, destructors, or virtual methods
-    size_t  mLoopStart;
-    size_t  mLoopEnd;
-    int     mLoopCount;
+
+    // These fields should both be size_t, but since they are located in shared memory we
+    // force to 32-bit.  The client and server may have different typedefs for size_t.
+    uint32_t    mLoopStart;
+    uint32_t    mLoopEnd;
+
+    int         mLoopCount;
 };
 
 }   // namespace android
